+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Updated tests for `ChartOption` class.
 - Make test classes final.
 
 ## 2.20.26 - 2025-10-04